+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.0
+
+* fix: remove SVGs for assets because they are generated as a SVG strings
+* feat: add `FeatherIconsTheme` Widget
+
 ## 0.1.0
 
 * chore: setup project